# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freelance Management API",
    "description": "Complete API documentation for freelance management system with products, orders, clients, and more",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/orders/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/orders/with-products": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create order with products",
        "description": "Create a new order with products, automatically calculating total price with tax, voucher discounts, and tips. Also reduces product quantities.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Order with products data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "required": {
                  "type": "array",
                  "example": [
                    "products"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "client_id": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "example": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    },
                    "products": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "required": {
                              "type": "array",
                              "example": [
                                "product_id",
                                "quantity",
                                "price"
                              ],
                              "items": {
                                "type": "string"
                              }
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "product_id": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "quantity": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 2
                                    }
                                  }
                                },
                                "price": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 99.99
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "voucher_refrence": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "VOUCHER-123"
                        }
                      }
                    },
                    "payment_methoud": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "enum": {
                          "type": "array",
                          "example": [
                            "CASH",
                            "CARD"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "example": {
                          "type": "string",
                          "example": "CARD"
                        }
                      }
                    },
                    "tip": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "number"
                        },
                        "example": {
                          "type": "number",
                          "example": 15
                        }
                      }
                    }
                  }
                },
                "example": {
                  "type": "object",
                  "properties": {
                    "client_id": {
                      "type": "number",
                      "example": 1
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "product_id": {
                            "type": "number",
                            "example": 2
                          },
                          "quantity": {
                            "type": "number",
                            "example": 1
                          },
                          "price": {
                            "type": "number",
                            "example": 49.99
                          }
                        }
                      }
                    },
                    "voucher_refrence": {
                      "type": "string",
                      "example": "VOUCHER-123"
                    },
                    "payment_methoud": {
                      "type": "string",
                      "example": "CARD"
                    },
                    "tip": {
                      "type": "number",
                      "example": 15
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order with products created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "integer"
                            },
                            "example": {
                              "type": "number",
                              "example": 201
                            }
                          }
                        },
                        "message": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "example": {
                              "type": "string",
                              "example": "Order with products created successfully"
                            }
                          }
                        },
                        "data": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "order_id": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "client_id": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "reference": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "ORD-123456"
                                    }
                                  }
                                },
                                "total_price": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 264.97
                                    }
                                  }
                                },
                                "tip": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 15
                                    }
                                  }
                                },
                                "payment_methoud": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "CARD"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "integer"
                            },
                            "example": {
                              "type": "number",
                              "example": 201
                            }
                          }
                        },
                        "message": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "example": {
                              "type": "string",
                              "example": "Order with products created successfully"
                            }
                          }
                        },
                        "data": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "order_id": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "client_id": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "integer"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "reference": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "ORD-123456"
                                    }
                                  }
                                },
                                "total_price": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 264.97
                                    }
                                  }
                                },
                                "tip": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 15
                                    }
                                  }
                                },
                                "payment_methoud": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "CARD"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation failed or insufficient stock"
          },
          "404": {
            "description": "Client or products not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}